refactor(api): split session/graph endpoints into api/routes_graph.py
app.py was 1323 lines mixing the core chat-completions proxy path with
~15 auxiliary REST endpoints (graph stats/viz/history/reasoning/obsidian
exports, manual checkpoints, decision invalidation, resume). Moved the
latter into a new APIRouter module so app.py now holds only app setup,
shared singletons/helpers, and POST /v1/chat/completions (995 lines,
down from 1323).

routes_graph.py references app.py's singletons (_analytics, _cache,
_checkpoint_mgr, _get_graph_async, _check_rate_limit) via
`app_module.<name>` rather than importing them by value, preserving the
same lazy-module-reference pattern the test suite already relies on for
monkeypatching (e.g. `app_module._analytics`, `app_module._graph_cache`)
— so those tests needed no changes despite the endpoints moving files.

Pure code motion, no behavior changes — full suite (437 tests, including
the TestClient-based integration tests that exercise these exact routes
over real HTTP) passes unchanged.
