+++ tokenmizer/api/app.py
@@ -1,8 +1,19 @@
 """
 TokenMizer — main FastAPI application.
 
-OpenAI-compatible proxy: POST /v1/chat/completions plus session/graph
-management endpoints. See README API Reference.
+OpenAI-compatible proxy: POST /v1/chat/completions, plus app setup,
+singletons, and the shared helpers (_get_graph_async, _check_rate_limit,
+_update_graph, etc.) that both this file and routes_graph.py depend on.
+See README API Reference.
+
+Session/graph inspection, checkpoint, and decision-management endpoints
+(~15 routes) live in routes_graph.py — split out to keep this file
+focused on the core proxy path. That module is imported at the bottom
+of this file and references the singletons/helpers defined here via
+`app_module.<name>` rather than importing them by value, so tests that
+monkeypatch this module's attributes (e.g. `app_module._analytics`,
+`app_module._graph_cache`) keep working unchanged regardless of which
+file actually handles a given request.
 """
 from __future__ import annotations
 
@@ -984,340 +995,12 @@ async def dashboard():
 
 
 # ── Session / Graph endpoints ─────────────────────────────────────────────────
+# Graph inspection, checkpoint, and decision-management endpoints live in
+# routes_graph.py (split out to keep this file focused on the core proxy
+# path). Imported at the bottom of this module — by this point every
+# singleton/helper routes_graph.py references via `app_module.<name>`
+# (_analytics, _cache, _checkpoint_mgr, _get_graph_async, _check_rate_limit)
+# is already defined above.
+from tokenmizer.api.routes_graph import router as _graph_router  # noqa: E402
 
-@app.get("/api/stats", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def stats(session_id: Optional[str] = None):
-    return _analytics.summary()
-
-
-@app.get("/api/cache/stats", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def cache_stats():
-    stats = _cache.stats()
-    # Include preference context for completeness (was previously unused)
-    stats["preference_context"] = _cache._preference_store.to_system_context()
-    return stats
-
-
-@app.get("/api/graph/{session_id}/history", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph_history(session_id: str, at_time: float = 0.0, top_k: int = 12):
-    """
-    Query graph state at a specific Unix timestamp.
-    at_time=0.0 (default) returns current state (equivalent to /viz).
-    at_time=<unix_ts> returns which nodes were active at that point in time.
-
-    Useful for: debugging decision changes, audit trail, "what did we decide
-    at 2pm?" queries.
-    """
-    graph = await _get_graph_async(session_id)
-    if at_time == 0.0:
-        nodes = graph.query("", top_k=top_k)
-    else:
-        nodes = graph.query_at_time("", at_time=at_time, top_k=top_k)
-    return {
-        "session_id": session_id,
-        "at_time": at_time or None,
-        "nodes": [
-            {
-                "id": n.id, "label": n.label, "type": n.type.value,
-                "status": n.status.value, "importance": n.importance,
-                "valid_from": n.valid_from, "valid_until": n.valid_until or None,
-            }
-            for n in nodes
-        ],
-        "count": len(nodes),
-    }
-
-
-@app.get("/api/graph/{session_id}", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph(session_id: str):
-    graph = await _get_graph_async(session_id)
-    return graph.stats()
-
-
-@app.get("/api/graph/{session_id}/viz", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph_viz(session_id: str):
-    """
-    Return full graph as D3-compatible JSON for visualization.
-    {nodes: [...], edges: [...], meta: {...}}
-    Used by the dashboard Graph tab and any external viz tool.
-    """
-    graph = await _get_graph_async(session_id)
-    return graph.to_vis_json()
-
-
-@app.get("/api/graph/{session_id}/html", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph_html(session_id: str):
-    """Shareable standalone interactive graph — open in a browser, drag/zoom,
-    screenshot, share. Self-contained dark-theme D3 force layout."""
-    from tokenmizer.graph_memory.visualization import to_share_html
-    graph = await _get_graph_async(session_id)
-    return HTMLResponse(to_share_html(graph))
-
-
-@app.get("/api/ontology", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_ontology():
-    """The TokenMizer graph ontology: node/edge types with semantics and
-    the status state machine. Machine-readable — what the graph CAN contain
-    and which lifecycle transitions are legal."""
-    from tokenmizer.graph_memory.ontology import ontology_dict
-    return ontology_dict()
-
-
-@app.get("/api/graph/{session_id}/why", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph_why(session_id: str, q: str):
-    """Reasoning: trace the causal chain behind a decision. Matches decision
-    nodes containing `q`, walks the supersession chain in both directions,
-    and returns the old→new trail with trigger/reason/evidence per hop,
-    plus the currently active choice."""
-    from tokenmizer.graph_memory.reasoning import why
-    graph = await _get_graph_async(session_id)
-    return why(graph, q)
-
-
-@app.get("/api/graph/{session_id}/reasoning", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph_reasoning(session_id: str):
-    """Full reasoning view over session memory: active decisions, recent
-    changes, decision history grouped by topic, and an ontology-based
-    consistency audit (contradictions, missing/dangling transitions)."""
-    from tokenmizer.graph_memory.reasoning import summarize_reasoning
-    graph = await _get_graph_async(session_id)
-    return summarize_reasoning(graph)
-
-
-@app.get("/api/graph/{session_id}/obsidian", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_graph_obsidian(session_id: str):
-    """
-    Download graph as Obsidian Canvas (.canvas) file.
-    Save as <any-name>.canvas inside your Obsidian vault and open directly.
-    """
-    import json as _json
-
-    from fastapi.responses import Response as _Resp
-    graph = await _get_graph_async(session_id)
-    canvas = graph.to_obsidian_canvas()
-    filename = f"tokenmizer-{session_id[:12]}.canvas"
-    return _Resp(
-        content=_json.dumps(canvas, indent=2),
-        media_type="application/json",
-        headers={"Content-Disposition": f"attachment; filename={filename}"},
-    )
-
-
-@app.get("/api/graph/{session_id}/transitions", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_transitions(session_id: str):
-    """Full decision transition history — trigger, reason, evidence, confidence_delta."""
-    graph = await _get_graph_async(session_id)
-    return {
-        "session_id": session_id,
-        "transitions": [
-            {
-                "id": t.id,
-                "from_label": t.from_label,
-                "to_label": t.to_label,
-                "trigger": t.trigger,
-                "reason": t.reason,
-                "evidence": t.evidence,
-                "confidence_delta": t.confidence_delta,
-                "timestamp": t.timestamp,
-                "context_line": t.to_context_line(),
-            }
-            for t in graph.get_transitions()
-        ],
-        "count": len(graph.get_transitions()),
-    }
-
-
-
-@app.post("/api/checkpoint", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def create_manual_checkpoint(session_id: str):
-    """
-    Create a manual checkpoint for a session, snapshotting current graph
-    state. Used by `tokenmizer checkpoint <session-id>` (CLI) and the
-    `/tokenmizer:checkpoint` Claude Code skill.
-
-    FOUND DURING A FINAL ACCURACY PASS: this endpoint was referenced by
-    the README's API Reference table, cli.py's `checkpoint` command, AND
-    the Claude Code checkpoint skill (.claude-plugin/skills/checkpoint/
-    SKILL.md) — all three call `POST /api/checkpoint?session_id=...` —
-    but it was never actually implemented here. Every one of those three
-    callers would have gotten a 404 against the real running app. This
-    wasn't a documentation typo; it was a real, consistent gap across
-    three independent consumers that nothing caught because none of them
-    were exercised end-to-end during the original audit.
-
-    Design note: unlike the auto-checkpoint path in chat_completions(),
-    this has no live message history to extract from (a standalone HTTP
-    call has no conversation attached) — `CheckpointManager.create()` is
-    called with `messages=[]`, which is safe: extract_from_messages()
-    early-returns on an empty new-messages diff, and the checkpoint still
-    correctly snapshots whatever's ALREADY in the graph from prior chat
-    turns. Verified with a direct test before writing this (see
-    tests/unit/test_graph_persistence.py for the equivalent pattern).
-    """
-    try:
-        graph = await _get_graph_async(session_id)
-        ckpt = _checkpoint_mgr.create(
-            session_id=session_id,
-            messages=[],
-            graph=graph,
-            context_pct=0.0,
-            trigger="manual",
-        )
-        return {
-            "checkpoint_id": ckpt.checkpoint_id,
-            "session_id": session_id,
-            "node_count": len(ckpt.graph_snapshot.get("nodes", [])),
-            "resume_tokens": ckpt.resume_tokens,
-            "resume_standard": ckpt.resume_standard,
-            "trigger": ckpt.trigger,
-        }
-    except Exception as e:
-        logger.error(f"Manual checkpoint failed for session {session_id}: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.get("/api/checkpoints/{session_id}", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def list_checkpoints(session_id: str):
-    return _checkpoint_mgr.list_checkpoints(session_id)
-
-
-_INVALIDATE_MIN_LABEL_LEN = 3  # matches validator.py's own noise-pattern floor (<=3 chars = noise)
-
-
-@app.post("/api/decision/invalidate", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def invalidate_decision(
-    session_id: str,
-    decision_label: Optional[str] = None,
-    node_id: Optional[str] = None,
-    reason: str = "",
-):
-    """
-    Mark a decision as INVALIDATED (red) — explicitly wrong or cancelled.
-    Use when a decision was made that turned out to be incorrect.
-    History is preserved; decision is flagged as a warning in future resumes.
-
-    Provide EITHER node_id (precise — targets exactly one node, e.g. an
-    id copied from /api/graph/{session_id}/viz or /transitions) OR
-    decision_label (fuzzy — word-boundary substring match against active
-    decision labels; may match more than one node, all of which are
-    returned in affected_nodes).
-
-    FIXED (TM-13): decision_label used to be matched with a raw substring
-    check (`label_lower in node.label.lower()`) and had no minimum
-    length — `decision_label=""` is a substring of every label, so an
-    empty (or accidentally-empty, e.g. a client bug that sends "") value
-    invalidated EVERY active decision in the session in one call. A short
-    label caused a milder version of the same problem: "sql" would match
-    "PostgreSQL" AND "SQLAlchemy" AND any future "MySQL" decision as a
-    side effect of literal substring containment, not because they're
-    actually related. Fixed by requiring a minimum length and matching on
-    a WORD-BOUNDARY substring instead of a raw one.
-    """
-    if node_id is None and (decision_label is None or
-                            len(decision_label.strip()) < _INVALIDATE_MIN_LABEL_LEN):
-        raise HTTPException(
-            status_code=400,
-            detail=(f"decision_label must be at least {_INVALIDATE_MIN_LABEL_LEN} "
-                    f"characters (a shorter value matches too many unrelated "
-                    f"decisions to safely invalidate) — or pass node_id for a "
-                    f"precise, single-node match."),
-        )
-    try:
-        import re
-
-        from tokenmizer.graph_memory.graph import NodeStatus, NodeType
-        graph = await _get_graph_async(session_id)
-
-        invalidated: list[dict] = []
-
-        if node_id is not None:
-            node = graph._nodes.get(node_id)
-            if (node is not None and node.type == NodeType.DECISION
-                    and node.status == NodeStatus.COMPLETED):
-                node.status = NodeStatus.INVALIDATED
-                node.summary = (
-                    f"Invalidated: {reason[:100]}" if reason else "Explicitly invalidated"
-                )
-                invalidated.append({"node_id": node_id, "label": node.label})
-        else:
-            label_lower = decision_label.lower().strip()
-            pattern = re.compile(r'\b' + re.escape(label_lower) + r'\b')
-            # Only ACTIVE (COMPLETED) decisions are eligible: matching
-            # across all statuses would let a label overwrite SUPERSEDED
-            # history nodes with INVALIDATED, destroying their
-            # supersession record. The response lists every affected node
-            # so multi-matches are visible.
-            for nid, node in graph._nodes.items():
-                if (node.type == NodeType.DECISION and
-                        node.status == NodeStatus.COMPLETED and
-                        pattern.search(node.label.lower())):
-                    node.status = NodeStatus.INVALIDATED
-                    node.summary = (
-                        f"Invalidated: {reason[:100]}" if reason else "Explicitly invalidated"
-                    )
-                    invalidated.append({"node_id": nid, "label": node.label})
-
-        if not invalidated:
-            target = node_id if node_id is not None else decision_label
-            raise HTTPException(
-                status_code=404,
-                detail=(f"No ACTIVE decision matching '{target}' found in "
-                        f"session '{session_id}' (superseded/archived decisions "
-                        f"are not invalidatable — they are already inactive)")
-            )
-        # direct node mutation above bypasses add_node's dirty-tracking —
-        # force=True is required here or this write is silently skipped
-        # (caught in a final accuracy pass; same class of bug the
-        # eviction path and prune() were already protected against).
-        # _persist() now returns bool (TM-12) — check it, since claiming
-        # "status": "invalidated" while the write actually failed is the
-        # same silent-data-loss pattern this whole audit is about.
-        if not graph._persist(force=True):
-            raise HTTPException(
-                status_code=500,
-                detail=(f"Decision(s) marked invalidated in memory, but the "
-                        f"write to disk FAILED for session '{session_id}' — "
-                        f"the change did not persist and will be lost on "
-                        f"restart or cache eviction. Retry the request."),
-            )
-        return {
-            "session_id": session_id,
-            "invalidated": node_id if node_id is not None else decision_label,
-            "affected_nodes": invalidated,
-            "reason": reason,
-            "status": "invalidated",
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Invalidate decision failed: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.get("/api/resume/{session_id}", dependencies=[Depends(verify_api_key), Depends(_check_rate_limit)])
-async def get_resume(session_id: str, level: str = "standard"):
-    """Get resume context for a session. level: critical | standard | full"""
-    try:
-        if level not in ("critical", "standard", "full"):
-            level = "standard"
-        ckpt = _checkpoint_mgr.get_latest(session_id)
-        if not ckpt:
-            raise HTTPException(status_code=404, detail="No checkpoint found for session")
-        resume_map = {
-            "critical": ckpt.resume_critical,
-            "standard": ckpt.resume_standard,
-            "full": ckpt.resume_full,
-        }
-        text = resume_map.get(level, ckpt.resume_standard)
-        return {
-            "session_id": session_id,
-            "checkpoint_id": ckpt.checkpoint_id,
-            "level": level,
-            "resume_context": text,
-            "token_count": count_tokens(text),
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Resume failed for {session_id}: {e}")
-        raise HTTPException(status_code=500, detail=f"Resume failed: {str(e)}")
+app.include_router(_graph_router)
